ghantparse/requirements.txt
geopy
httpx>=0.20.0
selenium-wire>=4.5.5
selenium>=4.0.0
imagehash
pillow
python-dateutil
colorama
beautifultable
termcolor
webdriver-manager==3.5.4
wayback
bs4
packaging
websockets
asgiref
redis
aiohttp==3.8.1